requirements.txt
# FastAPI and web server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3

# Google API integrations
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Slack API
slack-sdk==3.26.2

# Environment and configuration
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0

# Date/time handling
python-dateutil==2.8.2
pytz==2024.1

# HTTP requests
httpx==0.26.0
aiohttp==3.9.3

# Utilities
pyyaml==6.0.1
